just committing

+++ PathFinder.cpp
@@ -1,6 +1,6 @@
 #include "PathFinder.h"
 
-std::vector<Vec2d>
+std::vector<Vec2f>
 PathFinder::getPath(Scene* scene)
 {
 	cleanup();
@@ -9,7 +9,7 @@ PathFinder::getPath(Scene* scene)
 	int startY = (int)mScene->getStartPosition().y;
 	int startId = startY * mScene->getGrid().getMaxX() + startX;
 	mStartNode = new SearchNode(startX, startY, startId);
-	
+
 
 	int endX = (int)mScene->getEndPosition().x;
 	int endY = (int)mScene->getEndPosition().y;
@@ -25,7 +25,7 @@ PathFinder::getPath(Scene* scene)
 		return mPath;
 	}
 	else {
-		return std::vector<Vec2d>();
+		return std::vector<Vec2f>();
 	}
 }
 
@@ -91,7 +91,7 @@ PathFinder::searchNode(unsigned int x,unsigned int y, double movementCost, Searc
 
 	for (SearchNode* node : mOpenList)
 	{
-		if (id == node->id) 
+		if (id == node->id)
 		{
 			if (node->getExpectedCost() > newNode->getExpectedCost())
 			{
@@ -117,10 +117,10 @@ PathFinder::searchPath()
 		SearchNode* node = getNextNode();
 		if (node->id == mEndNode->id)
 		{
-			std::vector<Vec2d> inversePath;
+			std::vector<Vec2f> inversePath;
 			while (node != nullptr)
 			{
-				inversePath.push_back(Vec2d(node->x, node->y));
+				inversePath.push_back(Vec2f(node->x, node->y));
 				node = node->parent;
 			}
 
@@ -131,7 +131,7 @@ PathFinder::searchPath()
 		else {
 			int maxX = mScene->getGrid().getMaxX() - 1;
 			int maxY = mScene->getGrid().getMaxY() - 1;
-			
+
 			if (node->x < maxX)
 				searchNode(node->x + 1, node->y, node->movementCost + 1, node);
 			if (node->x > 0)
